Refactor component structure for easier testing
