--- app/controllers/test/components/mentoring/mentor_conversations_lists_controller.rb
@@ -1,3 +1,0 @@
-class Test::Components::Mentoring::MentorConversationsListsController < ApplicationController
-  def show; end
-end
