+++ app/controllers/test/components/mentoring/mentor_inboxes/conversations_controller.rb
@@ -1,4 +1,4 @@
-class Test::Components::Mentoring::MentorConversationsLists::ConversationsController < ApplicationController
+class Test::Components::Mentoring::MentorInboxes::ConversationsController < ApplicationController
   def index
     return head :internal_server_error if params[:state] == "Error" || params[:state] == "Loading"
