+++ app/controllers/test/components/mentoring/mentor_inboxes_controller.rb
@@ -1,0 +1,3 @@
+class Test::Components::Mentoring::MentorInboxesController < ApplicationController
+  def show; end
+end
