--- app/views/test/components/mentoring/mentor_conversations_lists/show.html.haml
@@ -1,7 +1,0 @@
-%h1 Mentor conversations list
-= form_tag nil, method: :get do
-  = label_tag :state
-  = select_tag :state, options_for_select(["Success", "Error", "Loading"], params[:state])
-  = submit_tag "Submit", name: nil
-%hr/
-= mentor_conversations_list(test_components_mentoring_mentor_conversations_list_conversations_path, test_components_mentoring_mentor_conversations_list_tracks_path, { state: params[:state] }, { retry: params[:state] === "Loading" })
