+++ app/views/test/components/mentoring/mentor_inboxes/show.html.haml
@@ -1,0 +1,7 @@
+%h1 Mentor inbox
+= form_tag nil, method: :get do
+  = label_tag :state
+  = select_tag :state, options_for_select(["Success", "Error", "Loading"], params[:state])
+  = submit_tag "Submit", name: nil
+%hr/
+= mentor_inbox(test_components_mentoring_mentor_inbox_conversations_path, test_components_mentoring_mentor_inbox_tracks_path, { state: params[:state] }, { retry: params[:state] === "Loading" })
