+++ test/javascript/components/mentoring/mentor_inbox/conversation_list/pagination.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { render, fireEvent } from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
-import { Pagination } from '../../../../../app/javascript/components/mentoring/mentor_conversations_list/pagination.jsx'
+import { Pagination } from '../../../../../../app/javascript/components/mentoring/mentor_inbox/conversation_list/pagination.jsx'
 
 test('clicking on "First" sets page to 1', () => {
   const setPage = jest.fn()
